third_party/blink/web_tests/external/wpt/css/css-display/parsing/tentative/display-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property display value 'inline grid-lanes'
  assert_equals: expected "inline grid-lanes" but got "inline-grid-lanes"
[FAIL] Property display value 'grid-lanes inline'
  assert_equals: expected "inline grid-lanes" but got "inline-grid-lanes"
Harness: the test ran to completion.
